pyproject.toml
[project]
name = "xperto"
version = "0.2.0"
description = "Your AI meeting assistant!"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.2.1",
    "ddgs>=9.6.1",
    "html2text>=2025.4.15",
    "pipecat-ai-tail>=0.0.1",
    "pipecat-ai[deepgram,elevenlabs,local-smart-turn,silero,speechmatics]>=0.0.84",
    "pyaudio>=0.2.14",
    "pydantic-settings>=2.9.1",
    "pydantic>=2.10.6",
    "pyyaml>=6.0.2",
    "textual>=3.1.1",
]

[project.scripts]
bot = "xperto.runner:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

[project.optional-dependencies]
daily = [
    "pipecat-ai[daily]>=0.0.84",
]
